Grab asv table from a tsv file and format to tibble

Grab asv table from a tsv file and format to tibble

Usage

asv_data_tsv(tsv)

Arguments

tsv

A character vector specifying the path to seqtab.tsv

Value

A tibble.

Examples

fpath <- system.file("extdata/tsv", "seqtab.tsv", package = "bubbler")
asv_data_tsv(fpath)
#> # A tibble: 10 × 21
#>    sample_id  ASV1  ASV2  ASV3   ASV4  ASV5  ASV6  ASV7  ASV8  ASV9 ASV10 ASV11
#>    <chr>     <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 Smp1          1     3   793      9    10  1267    29     0     0     1   290
#>  2 Smp2          0     3     0      0     1     0    91    11     0   317    26
#>  3 Smp3          0     3     0   4948   276     0     2     1    76     5    15
#>  4 Smp4         23     1     2      0     0    18    28     8   117  3464     1
#>  5 Smp5          7    23     2     51     3    19   198    27     2   231   145
#>  6 Smp6          0    15     3      2     0     6     0     0     2   116     1
#>  7 Smp7         10     0     1      3  1866     1    14     9     0     5   563
#>  8 Smp8         60     3     0      0     0     1     0    58  4025     0    78
#>  9 Smp9        173     0     0 123391     2    18    49     1    34    22     3
#> 10 Smp10         0   326     2      0     3     4   206     9    70     1    14
#> # ℹ 9 more variables: ASV12 <dbl>, ASV13 <dbl>, ASV14 <dbl>, ASV15 <dbl>,
#> #   ASV16 <dbl>, ASV17 <dbl>, ASV18 <dbl>, ASV19 <dbl>, ASV20 <dbl>