Pool taxa according to threshold

Applies a threshold and pools any taxa below this threshold, across samples and optionally, across a variable.

Usage

pool_taxa(
  rel_abund_tab,
  threshold,
  n_taxa,
  keep_metadata = FALSE,
  label = TRUE
)

Arguments

rel_abund_tab

A rel_abund table in tibble format.

threshold

A numeric vector for the threshold.

Value

A tibble.

Examples

rel_abund_phy(physeq1) %>% pool_taxa()
#> # A tibble: 30 × 3
#>    sample_id taxon          rel_abund
#>    <chr>     <glue>             <dbl>
#>  1 Smp1      < 0.0193%         0.0432
#>  2 Smp1      Actinobacteria    0.0360
#>  3 Smp1      Firmicutes        0.0216
#>  4 Smp10     < 0.0193%         0.0283
#>  5 Smp10     Actinobacteria    0.0218
#>  6 Smp10     Firmicutes        0.0306
#>  7 Smp2      < 0.0193%         0.0443
#>  8 Smp2      Actinobacteria    0.0428
#>  9 Smp2      Firmicutes        0.0264
#> 10 Smp3      < 0.0193%         0.0339
#> # ℹ 20 more rows