Generate a relative abundance table in tibble format from tsv files.

Generate a relative abundance table in tibble format from tsv files.

Usage

rel_abund_tsv(
  asv,
  taxa_data = NULL,
  meta_data = NULL,
  taxa_level = "Phylum",
  var = NULL
)

Arguments

asv

A tsv file path containing an asv table.

taxa_data

A tsv file path containing a taxa table.

meta_data

A tsv file path containing a meta_data table.

taxa_level

A character value specifying the taxa level from Domain to species.

var

A character value of a variable to sum by.

Value

A tibble.

Examples

asv <- system.file("extdata/tsv", "seqtab.tsv", package = "bubbler")
taxa <- system.file("extdata/tsv", "taxa.tsv", package = "bubbler")
meta_data <- system.file("extdata/tsv", "metadata.tsv", package = "bubbler")
rel_abund_tsv(asv, taxa, meta_data)
#> # A tibble: 200 × 8
#>    sample_id asv   level  taxon         rel_abund Depth Carbon_source Date      
#>    <chr>     <chr> <chr>  <chr>             <dbl> <dbl> <chr>         <date>    
#>  1 Smp1      ASV1  Phylum Actinomyceto…   6.09e-6    15 Hexadecane    2021-01-21
#>  2 Smp1      ASV2  Phylum Bacillota       1.83e-5    15 Hexadecane    2021-01-21
#>  3 Smp1      ASV3  Phylum Bacillota       4.83e-3    15 Hexadecane    2021-01-21
#>  4 Smp1      ASV4  Phylum Pseudomonado…   5.48e-5    15 Hexadecane    2021-01-21
#>  5 Smp1      ASV5  Phylum Pseudomonado…   6.09e-5    15 Hexadecane    2021-01-21
#>  6 Smp1      ASV6  Phylum Pseudomonado…   7.71e-3    15 Hexadecane    2021-01-21
#>  7 Smp1      ASV7  Phylum Bacteroidota    1.76e-4    15 Hexadecane    2021-01-21
#>  8 Smp1      ASV8  Phylum Bacillota       0          15 Hexadecane    2021-01-21
#>  9 Smp1      ASV9  Phylum Pseudomonado…   0          15 Hexadecane    2021-01-21
#> 10 Smp1      ASV10 Phylum Pseudomonado…   6.09e-6    15 Hexadecane    2021-01-21
#> # ℹ 190 more rows