Grab taxa table from a physeq object and format to tibble

Grab taxa table from a physeq object and format to tibble

Usage

taxa_data_phy(phy)

Arguments

phy

A phyloseq object containing tax_table

Value

A tibble.

Examples

taxa_data_phy(physeq)
#> # A tibble: 20 × 7
#>    asv   Domain   Phylum           Class               Order        Family Genus
#>    <chr> <chr>    <chr>            <chr>               <chr>        <chr>  <chr>
#>  1 ASV1  Bacteria Pseudomonadota   Alphaproteobacteria Rhodospiril… Azosp… Azos…
#>  2 ASV2  Bacteria Spirochaetota    Spirochaetia        Spirochaeta… Spiro… Alka…
#>  3 ASV3  Bacteria Pseudomonadota   Alphaproteobacteria Hyphomicrob… Phyll… Meso…
#>  4 ASV4  Bacteria Pseudomonadota   Deltaproteobacteria Desulfuromo… Geoba… Geom…
#>  5 ASV5  Bacteria Actinomycetota   Actinobacteria      Streptospor… Strep… Nono…
#>  6 ASV6  Bacteria Actinomycetota   Actinobacteria      Streptomyce… Strep… Stre…
#>  7 ASV7  Bacteria Pseudomonadota   Alphaproteobacteria Hyphomicrob… Methy… Meth…
#>  8 ASV8  Bacteria Bacillota        Tissierellia        Tissierella… Gotts… Gott…
#>  9 ASV9  Bacteria Pseudomonadota   Alphaproteobacteria Sphingomona… Sphin… Sphi…
#> 10 ASV10 Bacteria Bacteroidota     Flavobacteriia      Flavobacter… Flavo… Flav…
#> 11 ASV11 Bacteria Bacteroidota     Bacteroidia         Bacteroidal… Willi… Acet…
#> 12 ASV12 Bacteria Pseudomonadota   Betaproteobacteria  Burkholderi… Comam… Hydr…
#> 13 ASV13 Bacteria Bacteroidota     Bacteroidia         Bacteroidal… Prevo… Prev…
#> 14 ASV14 Bacteria Pseudomonadota   Deltaproteobacteria Desulfuromo… Desul… Desu…
#> 15 ASV15 Bacteria Pseudomonadota   Gammaproteobacteria Pseudomonad… Morax… Psyc…
#> 16 ASV16 Bacteria Deferribacterota Deferribacteres     Deferribact… Geovi… Geov…
#> 17 ASV17 Bacteria Actinomycetota   Actinobacteria      Micrococcal… Demeq… Deme…
#> 18 ASV18 Bacteria Actinomycetota   Actinobacteria      Mycobacteri… Gordo… Gord…
#> 19 ASV19 Bacteria Actinomycetota   Actinobacteria      Mycobacteri… Coryn… Cory…
#> 20 ASV20 Bacteria Bacillota        Bacilli             Lactobacill… Strep… Stre…