Grab taxa table from a tsv file and format to tibble

Grab taxa table from a tsv file and format to tibble

Usage

taxa_data_tsv(tsv)

Arguments

tsv

A character vector specifying the path to seqtab.tsv

Value

A tibble.

Examples

fpath <- system.file("extdata/tsv", "taxa.tsv", package = "bubbler")
taxa_data_tsv(fpath)
#> # A tibble: 20 × 7
#>    asv   Domain   Phylum         Class               Order          Family Genus
#>    <chr> <chr>    <chr>          <chr>               <chr>          <chr>  <chr>
#>  1 ASV1  Bacteria Actinomycetota Actinobacteria      Streptomyceta… Strep… Stre…
#>  2 ASV2  Bacteria Bacillota      Bacilli             Caryophanales  Alicy… Alic…
#>  3 ASV3  Bacteria Bacillota      Bacilli             Caryophanales  Staph… Stap…
#>  4 ASV4  Bacteria Pseudomonadota Gammaproteobacteria Alteromonadal… Alter… Mari…
#>  5 ASV5  Bacteria Pseudomonadota Gammaproteobacteria Pseudomonadal… Pseud… Pseu…
#>  6 ASV6  Bacteria Pseudomonadota Gammaproteobacteria Vibrionales    Vibri… Alii…
#>  7 ASV7  Bacteria Bacteroidota   Flavobacteriia      Flavobacteria… Flavo… Psyc…
#>  8 ASV8  Bacteria Bacillota      Bacilli             Lactobacillal… Lacto… Lact…
#>  9 ASV9  Bacteria Pseudomonadota Alphaproteobacteria Rickettsiales  Ricke… Rick…
#> 10 ASV10 Bacteria Pseudomonadota Gammaproteobacteria Pseudomonadal… Morax… Mora…
#> 11 ASV11 Bacteria Pseudomonadota Alphaproteobacteria Caulobacteral… Maric… Mari…
#> 12 ASV12 Bacteria Pseudomonadota Gammaproteobacteria Nevskiales     Nevsk… Sini…
#> 13 ASV13 Bacteria Bacillota      Bacilli             Lactobacillal… Lacto… Comp…
#> 14 ASV14 Bacteria Bacteroidota   Cytophagia          Cytophagales   Cyclo… Algo…
#> 15 ASV15 Bacteria Pseudomonadota Gammaproteobacteria Xanthomonadal… Xanth… Lyso…
#> 16 ASV16 Bacteria Bacillota      Bacilli             Caryophanales  Bacil… Anae…
#> 17 ASV17 Bacteria Bacillota      Clostridia          Eubacteriales  Clost… Clos…
#> 18 ASV18 Bacteria Pseudomonadota Gammaproteobacteria Enterobactera… Yersi… Serr…
#> 19 ASV19 Bacteria Pseudomonadota Alphaproteobacteria Rhodobacteral… Roseo… Rhod…
#> 20 ASV20 Bacteria Synergistota   Synergistia         Synergistales  Dethi… Rari…