Grab asv table from a physeq object and format to tibble

Grab asv table from a physeq object and format to tibble

Usage

asv_data_phy(phy)

Arguments

phy

A phyloseq object containing tax_table

Value

A tibble.

Examples

asv_data_phy(physeq1)
#> # A tibble: 10 × 11
#>    sample_id  ASV1  ASV2  ASV3  ASV4  ASV5  ASV6  ASV7  ASV8  ASV9 ASV10
#>    <chr>     <int> <int> <int> <int> <int> <int> <int> <int> <int> <int>
#>  1 Smp1         10    95    65    32    50    62    78    70    10    13
#>  2 Smp2         26    14    95    44    51    57    83    24    55    97
#>  3 Smp3         36    19    26    55    28    94    74     1    60    64
#>  4 Smp4         25    63    73    91    22    53     3    57    64     8
#>  5 Smp5         87    48    75    21    28    19    13    58    49    49
#>  6 Smp6         66    80     6    85    52    56     5    49    46    55
#>  7 Smp7         77    41     6    50    70    81    93    35    66    28
#>  8 Smp8         77    74     7    20    87    89    35    34    79    40
#>  9 Smp9         52    60    40    63    23    14    15    15    73    84
#> 10 Smp10        97    35    10    43    32     7    25     4    75    60